Overhauled and cleaned-up normalization approach.
