+++ languages/emf-metamodel/src/main/java/de/jplag/emf/MetamodelTokenType.java
@@ -16,14 +16,12 @@ public enum MetamodelTokenType implements TokenType {
     ENUM_END(ENUM),
     ENUM_LITERAL("EEnumLiteral"),
     OPERATION("EOperation"),
-    OPERATION_END(OPERATION),
     REFERENCE("EReference"),
     REFERENCE_MULT("EReference (multi-valued)"),
     ATTRIBUTE("EAttribute"),
     PARAMETER("EParameter"),
     INTERFACE("EInterface"),
     INTERFACE_END(INTERFACE),
-    SUPER_TYPE("ESuperType"),
     ID_ATTRIBUTE("EAttribute (ID)"),
     CONTAINMENT("EReference (Containment)"),
     CONTAINMENT_MULT("EReference (Containment, multi-valued)"),
