+++ languages/emf-metamodel/src/main/java/de/jplag/emf/normalization/TokenExtractionRules.java
@@ -1,4 +1,4 @@
-package de.jplag.emf.parser;
+package de.jplag.emf.normalization;
 
 import java.util.List;
 
@@ -17,8 +17,24 @@ import org.eclipse.emf.ecore.ETypeParameter;
 
 import de.jplag.emf.MetamodelTokenType;
 
-public class TokenExtractionRules {
+/**
+ * Utility class for the token extraction rules for the normalization.
+ */
+public final class TokenExtractionRules {
 
+    private TokenExtractionRules() {
+        // private constructor for non-instantiability.
+    }
+
+    /*
+     * TODO TS: These rules are currently duplicated from the token generator, if these rules converge we need to merge
+     * them. Furthermore they could also be moved to the token types themselves?
+     */
+    /**
+     * Returns the corresponding the token type for a model element.
+     * @param modelElement is the model element.
+     * @return the token type or null if no token is extracted for that element.
+     */
     public static MetamodelTokenType element2Token(EObject modelElement) {
         if (modelElement instanceof EAnnotation) {
             return MetamodelTokenType.ANNOTATION;
@@ -28,6 +44,7 @@ public class TokenExtractionRules {
             } else {
                 return MetamodelTokenType.ATTRIBUTE;
             }
+            // TODO TS: Decision for EAttribute?
             // return MetamodelTokenType.ATTRIBUTE;
         } else if (modelElement instanceof EClass eClass) {
             if (eClass.isInterface()) {
@@ -37,7 +54,6 @@ public class TokenExtractionRules {
             } else {
                 return MetamodelTokenType.CLASS;
             }
-            // return MetamodelTokenType.CLASS;
         } else if (modelElement instanceof EEnum) {
             return MetamodelTokenType.ENUM;
         } else if (modelElement instanceof EDataType) {
@@ -52,11 +68,24 @@ public class TokenExtractionRules {
             return MetamodelTokenType.PARAMETER;
         } else if (modelElement instanceof EReference eReference) {
             if (eReference.isContainment()) {
-                return MetamodelTokenType.CONTAINMENT;
+                if (eReference.getUpperBound() == 1) {
+                    return MetamodelTokenType.CONTAINMENT;
+                } else {
+                    return MetamodelTokenType.CONTAINMENT_MULT;
+                }
             } else {
-                return MetamodelTokenType.REFERENCE;
+                if (eReference.getUpperBound() == 1) {
+                    return MetamodelTokenType.REFERENCE;
+                } else {
+                    return MetamodelTokenType.REFERENCE_MULT;
+                }
             }
+            // TODO TS: Decision for EReference?
+            // if (eReference.isContainment()) {
+            // return MetamodelTokenType.CONTAINMENT;
+            // } else {
             // return MetamodelTokenType.REFERENCE;
+            // }
         } else if (modelElement instanceof ETypeParameter) {
             return MetamodelTokenType.TYPE_PARAMETER;
         }
@@ -64,6 +93,12 @@ public class TokenExtractionRules {
         return null;
     }
 
+    /**
+     * Returns the corresponding the token types for a list of model elements. See
+     * {@link TokenExtractionRules#element2Token(EObject)}.
+     * @param modelElements contains the model elements.
+     * @return the list of corresponding token types, might contain less entries than elements.
+     */
     public static List<MetamodelTokenType> elements2Tokens(List<EObject> modelElements) {
         return modelElements.stream().map(it -> element2Token(it)).filter(it -> it != null).toList();
     }
