+++ languages/emf-metamodel/src/main/java/de/jplag/emf/parser/MetamodelTokenGenerator.java
@@ -51,8 +51,7 @@ public class MetamodelTokenGenerator extends AbstractMetamodelVisitor {
      * Creates the visitor.
      * @param parser is the parser which receives the generated tokens.
      */
-    public MetamodelTokenGenerator(EcoreParser parser) {
-        super(true);
+    public MetamodelTokenGenerator(EcoreParser parser) { // TODO TS: Enable sorting again
         this.parser = parser;
     }
